COM-3633: display header
