+++ src/core/components/courses/TraineeTable.vue
@@ -1,12 +1,11 @@
 <template>
-  <ni-responsive-table :data="trainees" :columns="traineeColumns" v-model:pagination="traineePagination"
-    :hide-header="hideHeader" separator="none" :loading="loading" :class="tableClass"
-    :visible-columns="traineeVisibleColumns">
+  <ni-responsive-table :data="trainees" :columns="columns" v-model:pagination="traineePagination" separator="none"
+    :loading="loading" :class="tableClass" :visible-columns="visibleColumns">
     <template #header="{ props }">
-      <q-tr :props="props">
+      <q-tr :props="props" :class="[{ 'hide-header': hideHeader }]">
         <q-th v-for="col in props.cols" :key="col.name" :props="props" :style="col.style"
           :class="[{ 'table-actions-responsive': col.name === 'actions' }]">
-          {{ col.label }}
+          {{ hideHeader ? '' : col.label }}
         </q-th>
       </q-tr>
     </template>
@@ -45,7 +44,7 @@ import get from 'lodash/get';
 import pick from 'lodash/pick';
 import omit from 'lodash/omit';
 import { defineAbilitiesForCourse } from '@helpers/ability';
-import { formatPhone, formatPhoneForPayload } from '@helpers/utils';
+import { formatPhoneForPayload } from '@helpers/utils';
 import Courses from '@api/Courses';
 import Users from '@api/Users';
 import Button from '@components/Button';
@@ -62,6 +61,8 @@ export default {
     hideHeader: { type: Boolean, default: false },
     loading: { type: Boolean, default: false },
     tableClass: { type: String, default: () => '' },
+    columns: { type: Array, default: () => [] },
+    visibleColumns: { type: Array, default: () => [] },
   },
   components: {
     'ni-button': Button,
@@ -81,47 +82,6 @@ export default {
 
     const course = computed(() => $store.state.course.course);
 
-    const traineeColumns = computed(() => [
-      {
-        name: 'firstname',
-        label: 'Prénom',
-        align: 'left',
-        field: row => get(row, 'identity.firstname') || '',
-        classes: 'text-capitalize',
-      },
-      {
-        name: 'lastname',
-        label: 'Nom',
-        align: 'left',
-        field: row => get(row, 'identity.lastname') || '',
-        classes: 'text-capitalize',
-      },
-      { name: 'email', label: 'Email', align: 'left', field: row => get(row, 'local.email') || '', classes: 'email' },
-      {
-        name: 'phone',
-        label: 'Téléphone',
-        align: 'left',
-        field: row => get(row, 'contact.phone') || '',
-        format: formatPhone,
-      },
-      {
-        name: 'connectionInfos',
-        label: 'Code de connexion à l\'app',
-        field: 'loginCode',
-        align: 'center',
-      },
-      ...course.value.hasCertifyingTest
-        ? [{
-          name: 'certification',
-          label: 'Certification',
-          field: row => get(course.value, 'certifiedTrainees', []).includes(row._id),
-          format: value => (value ? 'Oui' : 'Non'),
-          align: 'center',
-        }]
-        : [],
-      { name: 'actions', label: '', align: 'right', field: '' },
-    ]);
-
     const traineeRules = {
       editedTrainee: {
         identity: { lastname: { required } },
@@ -137,12 +97,6 @@ export default {
       const ability = defineAbilitiesForCourse(pick(loggedUser.value, ['role']));
 
       return ability.can('update', subject('Course', course.value), 'trainees');
-    });
-
-    const traineeVisibleColumns = computed(() => {
-      const col = ['firstname', 'lastname', 'email', 'phone', 'connectionInfos', 'certification'];
-
-      return canUpdateTrainees.value ? [...col, 'actions'] : col;
     });
 
     const openTraineeEditionModal = async (trainee) => {
@@ -203,7 +157,6 @@ export default {
 
     return {
       // Data
-      traineeColumns,
       traineePagination,
       traineeEditionModal,
       traineeModalLoading,
@@ -212,7 +165,6 @@ export default {
       traineeValidation,
       // Computed
       course,
-      traineeVisibleColumns,
       canUpdateTrainees,
       // Methods
       openTraineeEditionModal,
@@ -234,4 +186,7 @@ export default {
 .border
   @media screen and (min-width: 767px)
     border-top: 1px solid $copper-grey-200
+.hide-header
+  height: 0px
+  display: block
 </style>
